Rename data table head components for consistency
