+++ src/components/DataTable/index.ts
@@ -1,4 +1,4 @@
 export * from "./DataTable";
 export * from "./DataTableBody";
-export * from "./DataTableHeader";
-export * from "./DataTableHeaderRow";
+export * from "./DataTableHead";
+export * from "./DataTableHeadRow";
